Renamed DistributionParam to DistOrArrayParam
This is more descriptive of its actual function, and lets us have a
DistributionParam that just does distributions. Also moved these
params to `dists.py`.
