+++ nengo/params.py
@@ -3,7 +3,6 @@ import weakref
 
 import numpy as np
 
-from nengo.dists import Distribution
 from nengo.processes import StochasticProcess
 from nengo.utils.compat import is_integer, is_number, is_string
 from nengo.utils.numpy import compare
@@ -37,7 +36,7 @@ class Parameter(object):
         Whether this parameter accepts the value None. By default,
         parameters are not optional (i.e., cannot be set to ``None``).
     readonly : bool, optional
-        Whether the parameter can be set multiple times.
+        If true, the parameter can only be set once.
         By default, parameters can be set multiple times.
     """
     def __init__(self, default=Unconfigurable, optional=False, readonly=False):
@@ -198,34 +197,6 @@ class NdarrayParam(Parameter):
                 raise ValueError("shape[%d] should be %d (got %d)"
                                  % (i, desired, ndarray.shape[i]))
         return ndarray
-
-
-class DistributionParam(NdarrayParam):
-    """Can be a Distribution or samples from a distribution."""
-
-    def __init__(self, default=Unconfigurable, sample_shape=None,
-                 optional=False, readonly=False):
-        super(DistributionParam, self).__init__(
-            default, sample_shape, optional, readonly)
-
-    def validate(self, instance, dist):
-        if dist is not None and not isinstance(dist, Distribution):
-            dist = super(DistributionParam, self).validate(instance, dist)
-        return dist
-
-
-class StochasticProcessParam(Parameter):
-    """Can be a StochasticProcess."""
-
-    def validate(self, instance, process):
-        super(StochasticProcessParam, self).validate(instance, process)
-
-        if process is not None and not isinstance(process, StochasticProcess):
-            raise ValueError(
-                "Must be StochasticProcess (got type {0}).".format(
-                    process.__class__.__name__))
-
-        return process
 
 
 FunctionInfo = collections.namedtuple('FunctionInfo', ['function', 'size'])
